Widen the lost events count to uint64_t
The lost_events callback reports how far a consumer fell behind in
ring buffer words. The value is a difference of two uint64_t ring
positions but the callback takes an int, so it wraps once the gap
exceeds INT_MAX.

+++ otherlibs/runtime_events/caml/runtime_events_consumer.h
@@ -77,7 +77,7 @@ CAMLextern void caml_runtime_events_set_lifecycle(
 /* Set the lost events callback on the cursor */
 CAMLextern void caml_runtime_events_set_lost_events(
     struct caml_runtime_events_cursor *cursor,
-    int (*f)(int domain_id, void *callback_data, int lost_words));
+    int (*f)(int domain_id, void *callback_data, uint64_t lost_words));
 
 /* Set the user events callback on the cursor */
 CAMLextern void caml_runtime_events_set_user_events(
